# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Efosa Hackerbay Microservice",
    "description": "Stateless microservice",
    "contact": {}
  },
  "host": "localhost:8000",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/patch": {
      "patch": {
        "description": "This endpoint would apply specified JSON patch on the specified document.",
        "summary": "ApplyJSONpatch",
        "tags": [
          "Misc"
        ],
        "operationId": "ApplyJSONpatch",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApplyJSONpatchRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/auth": {
      "post": {
        "description": "Endpoint to authenticate user and receive a jwt token used to access protected endpoints",
        "summary": "AuthenticateUser",
        "tags": [
          "Misc"
        ],
        "operationId": "AuthenticateUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/thumbnail": {
      "post": {
        "description": "Returns a generated thumbnail from specified image",
        "summary": "Generatethumbnail",
        "tags": [
          "Misc"
        ],
        "operationId": "Generatethumbnail",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "imageUrl",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "definitions": {
    "ApplyJSONpatchRequest": {
      "title": "ApplyJSONpatchRequest",
      "type": "object",
      "properties": {
        "document": {
          "type": "object"
        },
        "operations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Operations"
          }
        }
      },
      "required": [
        "document",
        "operations"
      ]
    },
    "Operations": {
      "title": "Operations",
      "type": "object",
      "properties": {
        "op": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "value": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "op",
        "path"
      ]
    }
  },
  "tags": [
    {
      "name": "Misc",
      "description": ""
    }
  ]
}